backend/requirements.txt
flask
flask-cors
transformers
torch
Pillow
opencv-python
google-generativeai
